# packages/api-client/openapi.json
{
  "openapi": "3.2.0",
  "paths": {
    "/stocks/suggestion": {
      "get": {
        "operationId": "StocksController_suggestion",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 1,
              "example": "삼ㅈ",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "example": 10,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true,
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SuggestionDto"
                    },
                    "meta": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": ["success", "data"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["Stocks"]
      }
    },
    "/stocks": {
      "get": {
        "operationId": "StocksController_search",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 1,
              "example": "삼성",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true,
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/StockDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": ["success", "data"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["Stocks"]
      }
    },
    "/stocks/{symbol}": {
      "get": {
        "operationId": "StocksController_get",
        "parameters": [
          {
            "name": "symbol",
            "required": true,
            "in": "path",
            "description": "Listed stock symbol",
            "schema": {
              "pattern": "^[A-Z0-9]{1,9}$",
              "example": "005930",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true,
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/StockDto"
                    },
                    "meta": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": ["success", "data"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "The requested stock symbol is not supported by this service.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "const": false,
                          "example": false
                        },
                        "error": {
                          "type": "object",
                          "properties": {
                            "status": {
                              "type": "integer",
                              "minimum": -9007199254740991,
                              "maximum": 9007199254740991,
                              "example": 404,
                              "enum": [404]
                            },
                            "type": {
                              "type": "string",
                              "example": "stock.unsupported",
                              "enum": ["stock.unsupported"]
                            },
                            "message": {
                              "type": "string",
                              "example": "Unsupported stock symbol"
                            },
                            "data": {
                              "type": "object",
                              "properties": {
                                "symbol": {
                                  "type": "string",
                                  "example": "005930"
                                }
                              },
                              "required": ["symbol"],
                              "additionalProperties": false
                            }
                          },
                          "required": ["status", "type", "message", "data"],
                          "additionalProperties": false
                        }
                      },
                      "required": ["success", "error"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Stocks"]
      }
    },
...
  },
  "info": {
    "title": "Trading API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SuggestionDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "symbol": {
                  "type": "string",
                  "pattern": "^[A-Z0-9]{1,9}$",
                  "description": "Listed stock symbol",
                  "example": "005930"
                },
                "name": {
                  "type": "string",
                  "example": "삼성전자"
                },
                "marketName": {
                  "type": "string",
                  "example": "KOSPI"
                },
                "quotationMarket": {
                  "type": "string",
                  "enum": ["KRX", "NXT", "CONSOLIDATED"],
                  "example": "KRX"
                },
                "standardCode": {
                  "example": "KR7005930003",
                  "type": "string"
                },
                "securityGroupCode": {
                  "example": "ST",
                  "type": "string"
                },
                "rawEtpType": {
                  "example": "2",
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "preferredStockType": {
                  "example": "0",
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "productType": {
                  "example": "STOCK",
                  "type": "string",
                  "enum": [
                    "STOCK",
                    "PREFERRED",
                    "ETF",
                    "ETN",
                    "REIT",
                    "BENEFICIARY_CERTIFICATE",
                    "SPAC",
                    "OTHER"
                  ]
                },
                "isPreferred": {
                  "example": false,
                  "type": "boolean"
                },
                "isEtf": {
                  "example": false,
                  "type": "boolean"
                },
                "isEtn": {
                  "example": false,
                  "type": "boolean"
                },
                "isSpac": {
                  "example": false,
                  "type": "boolean"
                },
                "isReit": {
                  "example": false,
                  "type": "boolean"
                },
                "isTradingHalted": {
                  "example": false,
                  "type": "boolean"
                },
                "isUnderAdministration": {
                  "example": false,
                  "type": "boolean"
                },
                "isLowLiquidity": {
                  "example": false,
                  "type": "boolean"
                },
                "marketCap": {
                  "example": 447733691250000,
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "previousVolume": {
                  "example": 12345678,
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "listedDate": {
                  "example": "19750611",
                  "anyOf": [
                    {
                      "type": "string",
                      "pattern": "^\\d{8}$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "isKospi100": {
                  "example": true,
                  "type": "boolean"
                },
                "isKospi50": {
                  "example": true,
                  "type": "boolean"
                },
                "isKrx300": {
                  "example": true,
                  "type": "boolean"
                },
                "warningLevel": {
                  "example": "00",
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": ["symbol", "name", "marketName", "quotationMarket"]
            }
          },
          "hasMore": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["items", "hasMore"]
      },
      "StockDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "pattern": "^[A-Z0-9]{1,9}$",
            "description": "Listed stock symbol",
            "example": "005930"
          },
          "name": {
            "type": "string",
            "example": "삼성전자"
          },
          "marketName": {
            "type": "string",
            "example": "KOSPI"
          },
          "quotationMarket": {
            "type": "string",
            "enum": ["KRX", "NXT", "CONSOLIDATED"],
            "example": "KRX"
          },
          "standardCode": {
            "example": "KR7005930003",
            "type": "string"
          },
          "securityGroupCode": {
            "example": "ST",
            "type": "string"
          },
          "rawEtpType": {
            "example": "2",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "preferredStockType": {
            "example": "0",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "productType": {
            "example": "STOCK",
            "type": "string",
            "enum": [
              "STOCK",
              "PREFERRED",
              "ETF",
              "ETN",
              "REIT",
              "BENEFICIARY_CERTIFICATE",
              "SPAC",
              "OTHER"
            ]
          },
          "isPreferred": {
            "example": false,
            "type": "boolean"
          },
          "isEtf": {
            "example": false,
            "type": "boolean"
          },
          "isEtn": {
            "example": false,
            "type": "boolean"
          },
          "isSpac": {
            "example": false,
            "type": "boolean"
          },
          "isReit": {
            "example": false,
            "type": "boolean"
          },
          "isTradingHalted": {
            "example": false,
            "type": "boolean"
          },
          "isUnderAdministration": {
            "example": false,
            "type": "boolean"
          },
          "isLowLiquidity": {
            "example": false,
            "type": "boolean"
          },
          "marketCap": {
            "example": 447733691250000,
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "previousVolume": {
            "example": 12345678,
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "listedDate": {
            "example": "19750611",
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{8}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "isKospi100": {
            "example": true,
            "type": "boolean"
          },
          "isKospi50": {
            "example": true,
            "type": "boolean"
          },
          "isKrx300": {
            "example": true,
            "type": "boolean"
          },
          "warningLevel": {
            "example": "00",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["symbol", "name", "marketName", "quotationMarket"]
      },
...
    }
  }
}